+++ packages/react/src/index.tsx
@@ -7,5 +7,7 @@ export * from './components/TextInput'
 export * from './components/TextArea'
 export * from './components/Checkbox'
 export * from './components/MultiStep'
+export * from './components/Stats'
+export * from './components/Tag'
 
 export * from './styles'