+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @iapp-ui/react
 
+## 3.0.3
+
+### Patch Changes
+
+- change ColorsUI and And Components
+
 ## 3.0.2
 
 ### Patch Changes